# docs/apidog-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FundLedger API",
    "version": "0.1.0",
    "description": "Apidog-importable OpenAPI collection for the FundLedger multi-tenant SaaS API."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    },
    {
      "url": "https://api.fundledger.app",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/v1/auth/otp/request": {
      "post": {
        "tags": ["Auth"],
        "summary": "Request one-time login code",
        "security": [],
        "requestBody": { "$ref": "#/components/requestBodies/OtpRequest" },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/v1/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login with mobile and OTP",
        "security": [],
        "requestBody": { "$ref": "#/components/requestBodies/LoginRequest" },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "401": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/v1/auth/switch-project": {
      "post": {
        "tags": ["Auth"],
        "summary": "Set active project for current session",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SwitchProjectRequest" },
              "example": { "project_id": "prj_towerB" }
            }
          }
        },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "403": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/v1/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Current user, tenant, active project and role",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectIdHeader" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "401": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/v1/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Revoke current token",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      }
    },
...
    "/v1/me/dues": {
      "get": {
        "tags": ["Dues"],
        "summary": "Signed-in member dues",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectIdHeader" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      }
    },
    "/v1/me/summary": {
      "get": {
        "tags": ["Dues"],
        "summary": "Signed-in member contribution summary",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectIdHeader" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      }
    },
    "/v1/dues": {
      "get": {
        "tags": ["Dues"],
        "summary": "Staff due list",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ProjectIdHeader" },
          { "name": "status", "in": "query", "schema": { "$ref": "#/components/schemas/DueStatus" } }
        ],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      }
    },
...
    "/v1/expenses": {
      "post": {
        "tags": ["Expenses"],
        "summary": "Initiate expense",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectIdHeader" }],
        "requestBody": { "$ref": "#/components/requestBodies/ExpenseCreate" },
        "responses": { "201": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/Error" } }
      },
      "get": {
        "tags": ["Expenses"],
        "summary": "List expenses",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ProjectIdHeader" },
          { "name": "status", "in": "query", "schema": { "$ref": "#/components/schemas/ExpenseStatus" } }
        ],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      }
    },
...
    "/v1/dues/{id}/penalty": {
      "get": {
        "tags": ["Penalties"],
        "summary": "Penalty breakdown for one due",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectIdHeader" }, { "$ref": "#/components/parameters/IdPath" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "404": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/v1/dues/{id}/penalty/waive": {
      "post": {
        "tags": ["Penalties"],
        "summary": "Waive accrued penalty",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectIdHeader" }, { "$ref": "#/components/parameters/IdPath" }],
        "requestBody": { "$ref": "#/components/requestBodies/Reason" },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/Error" } }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ProvisioningApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Billing/provisioning secret configured via PROVISIONING_API_KEYS."
      },
      "ProvisioningAuthorization": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Alternative form: Authorization: ApiKey <key>."
      }
    },
    "parameters": {
      "ProjectIdHeader": {
        "name": "X-Project-Id",
        "in": "header",
        "required": false,
        "schema": { "type": "string" },
        "description": "Active project scope. If omitted, the session active project is used."
      },
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      }
    },
    "responses": {
      "Success": {
        "description": "Success response envelope",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/SuccessEnvelope" }
          }
        }
      },
      "Error": {
        "description": "Error response envelope",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      }
    },
    "requestBodies": {
      "OtpRequest": {
        "required": true,
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/OtpRequest" },
            "example": { "mobile": "+8801711553300" }
          }
        }
      },
      "LoginRequest": {
        "required": true,
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/LoginRequest" },
            "example": {
              "mobile": "+8801711553300",
              "otp": "123456",
              "tenant_slug": "green-valley"
            }
          }
        }
      },
...
      "ExpenseCreate": {
        "required": true,
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ExpenseCreate" },
            "example": {
              "title": "Rod & cement procurement",
              "amount": 950000,
              "category": "materials",
              "vendor": "BSRM Steel",
              "doc_file_id": "file_bsrm_inv"
            }
          }
        }
      },
...
      "Reason": {
        "required": true,
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ReasonRequest" },
            "example": { "reason": "Invalid proof document" }
          }
        }
      }
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "required": ["ok", "data"],
        "properties": {
          "ok": { "type": "boolean", "example": true },
          "data": { "type": "object", "additionalProperties": true },
          "meta": { "type": "object", "additionalProperties": true }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": ["ok", "error"],
        "properties": {
          "ok": { "type": "boolean", "example": false },
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": { "type": "string", "example": "VALIDATION" },
              "message": { "type": "string", "example": "Invalid request" },
              "fields": { "type": "object", "additionalProperties": true }
            }
          }
        }
      },
...
      "DueStatus": {
        "type": "string",
        "enum": ["upcoming", "due", "overdue", "partial", "paid", "waived"]
      },
...
      "ExpenseStatus": {
        "type": "string",
        "enum": ["draft", "pending", "approved", "paid", "rejected"]
      },
      "ExpenseCategory": {
        "type": "string",
        "enum": ["materials", "construction", "professional", "services", "survey"]
      },
...
      "OtpRequest": {
        "type": "object",
        "required": ["mobile"],
        "properties": { "mobile": { "type": "string", "minLength": 6 } }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["mobile"],
        "properties": {
          "mobile": { "type": "string", "minLength": 6 },
          "otp": { "type": "string" },
          "tenant_slug": { "type": "string" },
          "project_id": { "type": "string" }
        }
      },
      "SwitchProjectRequest": {
        "type": "object",
        "required": ["project_id"],
        "properties": { "project_id": { "type": "string" } }
      },
...
      "ReasonRequest": {
        "type": "object",
        "required": ["reason"],
        "properties": { "reason": { "type": "string", "minLength": 3 } }
      },
      "ExpenseCreate": {
        "type": "object",
        "required": ["title", "amount", "category"],
        "properties": {
          "title": { "type": "string" },
          "amount": { "type": "integer", "minimum": 1 },
          "category": { "$ref": "#/components/schemas/ExpenseCategory" },
          "vendor": { "type": "string" },
          "doc_file_id": { "type": "string" }
        }
      },
...
    }
  }
}